update to get stub launcher class method to initialize() and update docstring
