+++ tfx/experimental/pipeline_testing/stub_component_launcher_test.py
@@ -37,11 +37,6 @@ from tfx.utils import io_utils
 
 from ml_metadata.proto import metadata_store_pb2
 
-class StubComponentLauncher(stub_component_launcher.StubComponentLauncher):
-  """Responsible for launching stub executors in Beam.
-  """
-  pass
-
 class CustomStubExecutor(base_stub_executor.BaseStubExecutor):
   """Example of custom stub executor.
 
@@ -108,13 +103,12 @@ class StubComponentLauncherTest(tf.test.TestCase):
     component_map = \
         {self.component.id: CustomStubExecutor}
 
-    my_stub_launcher = \
-        StubComponentLauncher.get_stub_launcher_class(
-            test_data_dir=self.record_dir,
-            stubbed_component_ids=[],
-            stubbed_component_map=component_map)
+    stub_component_launcher.StubComponentLauncher.initialize(
+        test_data_dir=self.record_dir,
+        stubbed_component_ids=[],
+        stubbed_component_map=component_map)
 
-    launcher = my_stub_launcher.create(
+    launcher = stub_component_launcher.StubComponentLauncher.create(
         component=self.component,
         pipeline_info=self.pipeline_info,
         driver_args=self.driver_args,
@@ -139,13 +133,12 @@ class StubComponentLauncherTest(tf.test.TestCase):
     io_utils.write_string_file(record_file, 'hello world')
     component_ids = [self.component.id]
 
-    my_stub_launcher = \
-        StubComponentLauncher.get_stub_launcher_class(
-            test_data_dir=self.record_dir,
-            stubbed_component_ids=component_ids,
-            stubbed_component_map={})
+    stub_component_launcher.StubComponentLauncher.initialize(
+        test_data_dir=self.record_dir,
+        stubbed_component_ids=component_ids,
+        stubbed_component_map={})
 
-    launcher = my_stub_launcher.create(
+    launcher = stub_component_launcher.StubComponentLauncher.create(
         component=self.component,
         pipeline_info=self.pipeline_info,
         driver_args=self.driver_args,
@@ -168,13 +161,12 @@ class StubComponentLauncherTest(tf.test.TestCase):
     io_utils.write_string_file(
         os.path.join(self.input_dir, 'result.txt'), 'test')
 
-    my_stub_launcher = \
-        StubComponentLauncher.get_stub_launcher_class(
-            test_data_dir=self.record_dir,
-            stubbed_component_ids=[],
-            stubbed_component_map={})
+    stub_component_launcher.StubComponentLauncher.initialize(
+        test_data_dir=self.record_dir,
+        stubbed_component_ids=[],
+        stubbed_component_map={})
 
-    launcher = my_stub_launcher.create(
+    launcher = stub_component_launcher.StubComponentLauncher.create(
         component=self.component,
         pipeline_info=self.pipeline_info,
         driver_args=self.driver_args,
